# web/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Web service that allows to run tests for programms written in almost any language.\n\nSchemes: [https]",
    "title": "Language Agnostic TEsting API",
    "contact": {
      "url": "https://kee-reel.com/about-ru"
    },
    "version": "0.0.1"
  },
  "basePath": "/",
  "paths": {
...
    "/login": {
      "get": {
        "produces": [
          "application/json"
        ],
        "tags": [
          "login"
        ],
        "summary": "Get token for registered user.",
        "operationId": "getLogin",
        "parameters": [
          {
            "type": "string",
            "description": "User email",
            "name": "email",
            "in": "URL",
            "required": true
          },
          {
            "type": "string",
            "format": "password",
            "description": "User password",
            "name": "pass",
            "in": "URL",
            "required": true
          }
        ]
      }
    },
...
  },
...
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}